[SSW-2210] Paging workarounds and improvements (#263)
Added fallbacks for paging API error (invalid cursor).
In case the existing paging logic fails, we use workarounds such as fetch400Profiles.
Removed API calls for devices included in profiles,
this should help with rate limits in the long term.
To do this we use the UDIDs to check for profile validity,
instead of the opaque ID from the API.